seperate the transactions page into two form
