+++ src/features/transactions/components/New/New.scss
@@ -1,3 +1,21 @@
+.mainContainer {
+  background: $background-color;
+  padding: 0;
+  display: flex;
+  flex-direction: row;
+  padding: 0 $gutter-size;
+}
+
+.content {
+  min-width: 400px;
+  width: 55%;
+  margin: 0 auto;
+}
+
+.submit {
+  text-align: right;
+}
+
 .submitTable {
   border-spacing: 5px;
